requirements.txt
streamlit
PyPDF2
google_generativeai
langchain
pdfplumber
chromadb
faiss-cpu
langchain_google_genai
langchain-community